+++ libraries/base/GHC/Exception.hs
@@ -140,9 +140,18 @@ class (Typeable e, Show e) => Exception e where
     toException = SomeException
     fromException (SomeException e) = cast e
 
+    -- | Render this exception value in a human-friendly manner.
+    --
+    -- Default implementation: @'show'@.
+    --
+    -- /Since: 4.8.0.0/
+    displayException :: e -> String
+    displayException = show
+
 instance Exception SomeException where
     toException se = se
     fromException = Just
+    displayException (SomeException e) = displayException e
 
 -- | Throw an exception.  Exceptions may be thrown from purely
 -- functional code, but may only be caught within the 'IO' monad.