+++ libraries/base/changelog.md
@@ -104,6 +104,8 @@
 
   * Add `fillBytes` to `Foreign.Marshal.Utils`.
 
+  * Add new `displayException` method to `Exception` typeclass. (#9822)
+
 ## 4.7.0.1  *Jul 2014*
 
   * Bundled with GHC 7.8.3